refactor: move encoder_decoder and kv_cache from hf/ to winml/
These modules are WinML-specific infrastructure (inference wrappers
and KV cache classes), not HuggingFace model configs. Move alongside
composite_model.py and decoder_only.py.

- src/winml/modelkit/models/hf/encoder_decoder.py -> winml/
- src/winml/modelkit/models/hf/kv_cache.py -> winml/

Updated imports in mu2.py, qwen.py, t5.py (hf modules now use
..winml.kv_cache), decoder_only.py, encoder_decoder.py (lazy imports
now use .kv_cache), and tests/unit/export/test_io.py.

Verified: Mu2 e2e (6/6 queries correct), all 79 UTs pass.
