+++ src/winml/modelkit/models/hf/clip.py
@@ -74,18 +74,17 @@ CLIP_CONFIG = WinMLBuildConfig(
 # =============================================================================
 @register_onnx_overwrite("clip_text_model", "feature-extraction", library_name="transformers")
 class CLIPTextModelIOConfig(CLIPTextWithProjectionOnnxConfig):
-    """ONNX config for CLIPTextModelWithProjection from transformers.
+    """ONNX config for CLIP text models (both with and without projection).
 
-    Model: openai/clip-vit-base-patch32 (text encoder only)
-    model.config.model_type = "clip_text_model"
+    Handles two architectures that share model_type ``clip_text_model``:
 
-    Inputs:
-        - input_ids: {0: "batch_size", 1: "sequence_length"}
-        - attention_mask: {0: "batch_size", 1: "sequence_length"}
+    - **CLIPTextModelWithProjection** (standalone CLIP text encoder):
+      Outputs ``text_embeds`` + ``last_hidden_state``
+    - **CLIPTextModel** (e.g., Stable Diffusion text encoder):
+      Outputs ``last_hidden_state`` only
 
-    Outputs:
-        - text_embeds: {0: "batch_size"}
-        - last_hidden_state: {0: "batch_size", 1: "sequence_length"}
+    The ``outputs`` property auto-detects which variant to use based on
+    ``config.architectures``.
 
     Key difference from Optimum's default:
         - sequence_length = max_position_embeddings (77 for CLIP)
@@ -104,6 +103,29 @@ class CLIPTextModelIOConfig(CLIPTextWithProjectionOnnxConfig):
         return {
             "input_ids": {0: "batch_size", 1: "sequence_length"},
             "attention_mask": {0: "batch_size", 1: "sequence_length"},
+        }
+
+    @property
+    def outputs(self) -> dict[str, dict[int, str]]:
+        """Return output tensors, adapting to the model architecture.
+
+        CLIPTextModelWithProjection produces ``text_embeds`` (projected)
+        and ``last_hidden_state``.  CLIPTextModel (used in Stable Diffusion)
+        produces only ``last_hidden_state``.
+
+        Default (architectures unset) assumes projection model for backward
+        compatibility.
+        """
+        architectures = getattr(self._config, "architectures", None) or []
+        # CLIPTextModel (non-projection, e.g. SD text_encoder)
+        if "CLIPTextModel" in architectures and "CLIPTextModelWithProjection" not in architectures:
+            return {
+                "last_hidden_state": {0: "batch_size", 1: "sequence_length"},
+            }
+        # Default: projection model outputs (backward compatible)
+        return {
+            "text_embeds": {0: "batch_size"},
+            "last_hidden_state": {0: "batch_size", 1: "sequence_length"},
         }
